Add support wasm32 platform with simd128

+++ src/main.zig
@@ -51,7 +51,7 @@ fn bitonicSortSample() void {
 }
 
 fn vqsortSample() !void {
-    const IntType = u16;
+    const IntType = u32;
     var prnd = std.rand.DefaultPrng.init(83751737);
 
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
